start redoing structure and logic of logs. Refactor from multiple helper functions to single map function inline with component.
