+++ src/components/ProcessLogsCard/ProcessLogsCard.module.scss
@@ -1,0 +1,6 @@
+.card {
+  background-color: color(background, light);
+  border-radius: 0.5rem;
+  margin: 1rem;
+  padding: 0.25rem 0.5rem;
+}
