+++ src/components/ProcessLogsCard/ProcessLogsCard.tsx
@@ -3,6 +3,9 @@ import { useNavigate } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import { LogsCardProps } from '../../../types';
 
+import styles from './ProcessLogsCard.module.scss';
+import globalStyles from '../global.module.scss';
+
 /**
  * @module | ProcessLogsCard.tsx
  * @description | Provides container cards organized based on container status; clicking will allow user to view container-specific process logs
@@ -12,15 +15,14 @@ const ProcessLogsCard = (props: LogsCardProps): JSX.Element => {
   const navigate = useNavigate();
 
   return (
-    <div className="card w-96 glass">
-      <div className="card-body">
-        <h2 className="card-title">{props.container.Names}</h2>
-        <div className="divider py-1"></div>
-        <p className="text-xs">Container ID: {props.container.ID}</p>
-        <p className="text-xs">Container Status: {props.status}</p>
-        <div className="card-actions justify-end items-center">
+    <div className={styles.card}>
+      <div className={styles.test}>
+        <h2>{props.container.Names}</h2>
+        <p>Container ID: {props.container.ID}</p>
+        <p>Container Status: {props.status}</p>
+        <div>
           <button
-            className="btn btn-primary px-10 py-0 w-35 text-center font-bold text-primary-content text-sm"
+            className={globalStyles.buttonSmall}
             onClick={() => navigate(`/home/logTable/${props.container.ID}`)}
           >
             VIEW LOGS
