+++ src/components/ProcessLogsTable/ProcessLogsTable.module.scss
@@ -1,0 +1,3 @@
+.wrapper {
+  display: flex;
+}
