+++ src/components/main/index.ts
